Singlenton y otros metodos, falta COR para corroborar todo

+++ TallerIntegradorPatrones/src/ChainOfResponsibility/Manejador.java
@@ -8,18 +8,18 @@ package ChainOfResponsibility;
 public class Manejador
 {
     protected int monto;
-    protected int denominacion;
+    protected double denominacion;
 
-    public Manejador(int monto, int denominacion){
+    public Manejador(int monto, double denominacion){
         this.monto = monto; // Total de billetes
         this.denominacion = denominacion; // Valor de cada billete
     }
-
+    
     public int getMonto(){ return monto; }
-    public int getDenominacion(){ return denominacion; }
+    public double getDenominacion(){ return denominacion; }
     public void setMonto(int monto){ this.monto = monto; }
 
-    public boolean retirar(int monto){
+    public boolean retirar(double monto){
         // Implementar
         return false;
     }
